Rename of the files for the configuration to work.
