+++ test/index.js
@@ -11,7 +11,7 @@ require('./runner');
 require('./source_map');
 require('./validator');
 require('./rules/fields_have_descriptions');
-require('./rules/fields_are.camelcased');
+require('./rules/fields_are_camel_cased.js');
 require('./rules/types_have_descriptions');
 require('./rules/deprecations_have_a_reason');
 require('./rules/enum_values_sorted_alphabetically');
@@ -19,7 +19,7 @@ require('./rules/enum_values_all_caps');
 require('./rules/types_are_capitalized');
 require('./rules/defined_types_are_used');
 require('./rules/input_object_values_have_descriptions');
-require('./rules/input_object_values_are_camelcased');
+require('./rules/input_object_values_are_camel_cased');
 require('./rules/enum_values_have_descriptions');
 require('./formatters/json_formatter');
 require('./formatters/text_formatter');
